Improve pty reading and renderer synchronization
The pty read thread now runs the parser and directly updates the
terminal in the same thread. This obviates the need for a channel which
sends every char read from the pty; this is a huge performance boon.

Synchronization between the updater and the renderer is now achieved
with a PriorityMutex. Previously, an atomic bool was (poorly) used to
request the lock on terminal. The PriorityMutex is dead simple to use,
and it _Just Works_.

+++ src/io.rs
@@ -69,7 +69,7 @@ pub enum Utf8CharsError {
     /// of a byte sequence well-formed in UTF-8, but ends prematurely.
     ///
     /// Contains number of unused bytes
-    IncompleteUtf8(u8),
+    IncompleteUtf8(usize),
 
     /// Variant representing that an I/O error occurred.
     Io(Error),
